feat: ✨ Add a new docs website to share with stakeholders
The docs directory build into a Github Pages site. Some contents from our wiki and README file are in this site. We'll need to deprecate that content elsewhere once this site is complete. This contains all of our pertinent documentation to-date.

Some wiki and README content will need to be removed in favor of this docs site.